split movie runner and events
